refactor: 불필요한 로그 제거 및 오류 수정 1. 우승자, 랜덤숫자, 길이 등 확인하던 로그 제거
2. 우승자가 복수명일시 출력이 되지 않던 오류 수정
