+++ src/view/print-result.js
@@ -1,12 +1,15 @@
-import { state } from '..';
-import { LINE_BREAK, POSITION_CONDITION_NUMBER } from '../constants';
-import { addRandomPosition, decideWinner } from '../game-utils';
+import { state } from '../index.js';
+import {
+  LINE_BREAK,
+  POSITION_CONDITION_NUMBER,
+} from '../constants/constants.js';
+import { addRandomPosition, decideWinner } from '../utils/game-utils.js';
 
 export const printGameResult = function (racingCount) {
   addRandomPosition(racingCount);
   const resultSpan = document.querySelector('#result-span');
 
-  for (let i = 1; i < i <= racingCount; i++) {
+  for (let i = 1; i <= racingCount; i++) {
     for (const car of state.carArray) {
       resultSpan.innerHTML += `${LINE_BREAK} ${car.name}: `;
       printPostions(car, i);
@@ -28,6 +31,7 @@ const printPostions = function (car, round) {
 };
 
 export const printFinalWinner = function () {
+  decideWinner();
   const finalWinner = decideWinner().join(', ');
   const resultSpan = document.querySelector('#result-span');
   const racingWinners = document.querySelector('#racing-winners');
